Added CLA Group project_live attribute support
- Added CLA Group project_live attribute/column to allow PROD reporting
  of which repos are actually LIVE used by the community.
- Updated Python Data Model
- Updated GoLang Data Models
- Updated Project repository - moved helpers and projections to separate file
